requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==2.3.7
email-validator==2.1.0.post1
python-dotenv==1.0.0
WTForms==3.1.0 